// src/enums.ts
export enum ButtonVariant {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  UNDERLINED = 'underlined',
  MUTED = 'muted',
  DANGER = 'danger',
}

export enum ListType {
  WATCH_LIST = ' WATCH_LIST',
  FAVORITES = 'FAVORITES',
}

export enum PanelType {
  SEARCH = ' SEARCH',
  FILTERS = 'FILTERS',
}

export enum MovieTabType {
  POPULAR = 'popular',
  UPCOMING = 'upcoming',
  TOP_RATED = 'topRated',
  NOW_PLAYING = 'nowPlaying',
}